refactor: LottoConstant 의존성 줄이는 방향으로 패키지 변경
