+++ src/main/java/lotto/domain/Lotto.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import lotto.constant.LottoConstant;
+import lotto.domain.constant.LottoConstant;
 import lotto.domain.exception.ExceptionMessage;
 
 public class Lotto {
