+++ src/main/java/lotto/domain/LottoPurchase.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.constant.LottoConstant;
+import lotto.domain.constant.LottoConstant;
 import lotto.domain.exception.ExceptionMessage;
 import lotto.service.LottoGenerator;
