+++ src/main/java/lotto/domain/LottoWinningNumber.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.constant.LottoConstant;
+import lotto.domain.constant.LottoConstant;
 import lotto.domain.exception.ExceptionMessage;
 
 public class LottoWinningNumber {
