+++ src/main/java/lotto/domain/constant/LottoConstant.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.domain.constant;
 
 public class LottoConstant {
