+++ src/main/java/lotto/generator/RandomLottoGenerator.java
@@ -4,8 +4,8 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import lotto.constant.LottoConstant;
 import lotto.domain.Lotto;
+import lotto.domain.constant.LottoConstant;
 import lotto.service.LottoGenerator;
 
 public class RandomLottoGenerator implements LottoGenerator {
