+++ src/test/java/lotto/domain/LottoPurchaseTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.constant.LottoConstant;
+import lotto.domain.constant.LottoConstant;
 import lotto.generator.RandomLottoGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
