+++ src/test/java/lotto/generator/RandomLottoGeneratorTest.java
@@ -3,8 +3,8 @@ package lotto.generator;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.constant.LottoConstant;
 import lotto.domain.Lotto;
+import lotto.domain.constant.LottoConstant;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
